+++ pallas-multiplexer/src/agents.rs
@@ -116,3 +116,63 @@ impl<C: Channel> From<C> for ChannelBuffer<C> {
         ChannelBuffer::new(channel)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::collections::VecDeque;
+
+    use super::*;
+
+    impl Channel for VecDeque<Payload> {
+        fn enqueue_chunk(&mut self, chunk: Payload) -> Result<(), ChannelError> {
+            self.push_back(chunk);
+            Ok(())
+        }
+
+        fn dequeue_chunk(&mut self) -> Result<Payload, ChannelError> {
+            let chunk = self.pop_front().ok_or(ChannelError::NotConnected(None))?;
+            Ok(chunk)
+        }
+    }
+
+    #[test]
+    fn multiple_messages_in_same_payload() {
+        let mut input = Vec::new();
+        let in_part1 = (1u8, 2u8, 3u8);
+        let in_part2 = (6u8, 5u8, 4u8);
+
+        minicbor::encode(in_part1, &mut input).unwrap();
+        minicbor::encode(in_part2, &mut input).unwrap();
+
+        let mut channel = VecDeque::<Payload>::new();
+        channel.push_back(input);
+
+        let mut buf = ChannelBuffer::new(channel);
+
+        let out_part1 = buf.recv_full_msg::<(u8, u8, u8)>().unwrap();
+        let out_part2 = buf.recv_full_msg::<(u8, u8, u8)>().unwrap();
+
+        assert_eq!(in_part1, out_part1);
+        assert_eq!(in_part2, out_part2);
+    }
+
+    #[test]
+    fn fragmented_message_in_multiple_payloads() {
+        let mut input = Vec::new();
+        let msg = (11u8, 12u8, 13u8, 14u8, 15u8, 16u8, 17u8);
+        minicbor::encode(msg, &mut input).unwrap();
+
+        let mut channel = VecDeque::<Payload>::new();
+
+        while !input.is_empty() {
+            let chunk = Vec::from(input.drain(0..2).as_slice());
+            channel.push_back(chunk);
+        }
+
+        let mut buf = ChannelBuffer::new(channel);
+
+        let out_msg = buf.recv_full_msg::<(u8, u8, u8, u8, u8, u8, u8)>().unwrap();
+
+        assert_eq!(msg, out_msg);
+    }
+}
